Avoiding premature instantiation of certain templates
- factored out template specializations for tuple types
