+++ hpx/parallel/traits/detail/boost_simd/vector_pack_type.hpp
@@ -9,28 +9,47 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_DATAPAR_BOOST_SIMD)
-#include <hpx/util/tuple.hpp>
 
 #include <cstddef>
+#include <type_traits>
 
 #include <boost/simd.hpp>
 
 ///////////////////////////////////////////////////////////////////////////////
 namespace hpx { namespace parallel { namespace traits
 {
-    template <typename T,
-        std::size_t N = boost::simd::native_cardinal<T>::value,
-        typename Abi = boost::simd::abi_of_t<T, N> >
-    struct vector_pack_type
+    ///////////////////////////////////////////////////////////////////////////
+    namespace detail
     {
-        typedef boost::simd::pack<T, N, Abi> type;
-    };
+        template <typename T, std::size_t N, typename Abi>
+        struct vector_pack_type
+        {
+            typedef typename std::conditional<
+                    std::is_void<Abi>::value, boost::simd::abi_of_t<T, N>, Abi
+                >::type abi_type;
 
-    template <typename ... T, std::size_t N, typename Abi>
-    struct vector_pack_type<hpx::util::tuple<T...>, N, Abi>
-    {
-        typedef hpx::util::tuple<boost::simd::pack<T, N, Abi>...> type;
-    };
+            typedef boost::simd::pack<T, N, abi_type> type;
+        };
+
+        template <typename T, typename Abi>
+        struct vector_pack_type<T, 0, Abi>
+        {
+            static std::size_t const N = boost::simd::native_cardinal<T>::value;
+            typedef typename std::conditional<
+                    std::is_void<Abi>::value, boost::simd::abi_of_t<T, N>, Abi
+                >::type abi_type;
+
+            typedef boost::simd::pack<T, N, abi_type> type;
+        };
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+    // Avoid premature instantiation of boost::simd::native_cardinal and
+    // boost::simd::abi_of_t.
+    template <typename T, std::size_t N, typename Abi>
+    struct vector_pack_type
+      : detail::vector_pack_type<T, N, Abi>
+    {};
 }}}
 
 #endif
